Fixed a memory leak in basedclient.cpp within on_message() ERROR_DATA case, where the observable itself was not delete(d).
Fixed an incorrect erasure in the same case on m_observe_subs which was using the sub_id instead of the obs_id.

Improved the read_bytes_from_string function by including bounds checking and using bitwise shifts & ORs for the calculation so that it's more readable. Also made sure to use uint64 to match the method signature's return type.